src/requirements.txt
streamlit==1.36.0
torch==2.3.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
pillow==10.3.0
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
groq==0.9.0